cleaned up noise example

+++ examples/noise.rs
@@ -11,16 +11,16 @@ fn main() {
     let mut window = Window::new().unwrap();
 
     let seed = noise::Seed::new(0);
-    let mut z = 0i32;
+    let mut z = 0.0;
 
     while window.is_open() {
-        z += 1;
+        z += 1.0;
         let mut frame = window.frame(colors::RED);
 
         frame.draw_pixels(
             iter_2d(range(0, 256), range(0, 245)).map(|(x, y)| {
                 let (x, y) = (x as f32, y as f32);
-                let value = noise::simplectic3(&seed, &[x / DIV, y / DIV, z as f32 / DIV]);
+                let value = noise::simplectic3(&seed, &[x / DIV, y / DIV, z / DIV]);
                 let value = (value + 1.0) / 2.0;
 
                 ((x, y), [value, value, value])
